+++ serve/mlc_serve/engine/metrics_labels.py
@@ -1,0 +1,4 @@
+NUM_CACHE_EVICTONS = "num_cache_evictions"
+E2E_LATENCY = "e2e_latency"
+FIRST_TOKEN_LATENCY = "first_token_latency"
+KV_CACHE_UTILIZATION = "kv_cache_utilization"
